resolves issues from review
